Applied recent FL patches from Benjamin Williams <biwilliajsb@yahoo.com>.
1) There is a problem
in that floating frames are not currently supported
with !mRealTimeUpdates.  We wrote a patch to correct
this problem.  You will find it attached to this
posting.

Here is a description of what was changed and why.  In
cbBarDragPlugin::ShowHint(), SetBarState is called
every time mpCurPane changes.  This is correct.
However, this also happens every time, even if
mRealTimeUpdates is false.  The state should not be
changed during the drag operation if mRealTimeUpdates
is off.  This is corrected in this patch.  Code was
also added in cbBarDragPlugin::OnLButtonUp() to
actually do the state changing, only if
mRealTimeUpdates is off.  Normally, this is performed
in ShowHint if mRealTimeUpdates is on.

I also took the liberty of changing the drag cursor
back to an arrow.  This is the way it is in MS Visual
C++, and it looks way better.  The MoveWindow cursor
looks terrible, IMHO.  When FL gets the embedded
cursor code working, we can switch back to a hand or
something else.


2) This time we have
discovered a crash bug in FL.  The steps to reproduce this
bug are as follows:

1. Open up a pane, dock it.
2. Click the Expand button (arrow button) in the
   frame hints portion
3. Click it again to Collapse the pane.
4. Now click the close button of the pane. Crash.

We investigated what was causing this problem, and it turns
out that the button state never gets unset.  Once a button
is pressed, it is always pressed.  Thus, when the expand/
collapse button is pressed, and then the close button is
pressed, the pane will close, but then FL will also try to
expand or collapse the pane as well, because it thinks that
the expand/collapse button was set.