Applied patch [ 615769 ] generic wxListCtrl enhancements
implemented a couple of things missing from the generic
wxListCtrl class which make it more similar to the
Windows port

1) implemented wxLIST_FORMAT_RIGHT and wxLIST_FORMAT_CENTER

2) added an ellipsis (...) to the end of the string and
shortened the string to fit inside the allowed width

3) left a margin of 8 pixels between columns so that
the text is not shoved right together

Note this is my first patch to wxWindows. I did read
the patch howto and I think I got it right but please
feel free to bother me about any discrepancies. In
addition, I am not sure if the way I implemented the
new routine to handle the formatting is adequate
performance wise but it doesn't seem to slow things
down too badly on my machine... :-) In addition, if
this routine is accepted, it could also be used for the
header data as well -- thoughts?

Anthony Tuininga