Patch #1003174: Fixes crash when toolbar has a disabled control from Ian Brown.
  If you put a control (e.g. a wxChoice) on a toolbar and then
later disable that control, you will get a crash. This is because
the toolbar code always calls disable on the button widget for a
toolbartool. When the tool is a control, there is no button, and
you should call disable on the control instead.
This small patch implements this logic.