Fix signatures of various image handlers methods.
Use wxBitmapType instead of long for the type parameter and made SaveFile()
methods const to follow the base class changes.

Now that the signatures are the same as in the base class these handlers have
a chance to work again while their methods couldn't be called at all before
this change.