Implement wxGetActiveWindow() for wxX11.
Now that wxGetActiveWindow() is used when showing modal dialogs, it became
critical to implement it as otherwise an assert happens every time a dialog is
shown and because creating the assert dialog itself results in another (same)
assert, this immediately results in a crash.

So implement it even if in a very trivial (but hopefully not too incorrect)
way.