Simplify timezone-related code and fix some minor bugs in it.
Try to make the chain of preprocessor checks for different ways of getting
time zone from the CRT more clear.

Also call _tzset() for all MSVC versions, not just MSVC8+ (closes #12700). We
should probably call tzset() for the other compilers too, in fact.

And multiply the timezone returned from ftime() by 60 as it's supposed to be
in minutes and not seconds as needed.